refactor: unify PMID/linking utilities for alteration cards/report
